+++ fake-koji/src/test/java/org/fakekoji/jobmanager/JenkinsJobUpdaterTest.java
@@ -24,10 +24,19 @@
 package org.fakekoji.jobmanager;
 
 import java.io.IOException;
-import java.util.Arrays;
+
+
+import org.fakekoji.DataGenerator;
+import org.fakekoji.core.AccessibleSettings;
+import org.fakekoji.jobmanager.model.JobUpdateResults;
+import org.fakekoji.jobmanager.project.JDKProjectManager;
+import org.fakekoji.jobmanager.project.JDKTestProjectManager;
+import org.fakekoji.storage.StorageException;
 import org.junit.Assert;
-import org.junit.Assume;
+import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 
 /**
  * Warning - reaming check have missing check on content!
@@ -37,6 +46,22 @@ import org.junit.Test;
 public class JenkinsJobUpdaterTest {
 
     boolean wasFinallyRun = false;
+
+    @Rule
+    public final TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+    private AccessibleSettings settings;
+    private DataGenerator.FolderHolder folderHolder;
+    private JobUpdater jobUpdater;
+    private ConfigManager cm;
+
+    @Before
+    public void setup() throws IOException, StorageException {
+        folderHolder = DataGenerator.initFolders(temporaryFolder);
+        cm = ConfigManager.create(folderHolder.configsRoot.getAbsolutePath());
+        settings = DataGenerator.getSettings(folderHolder);
+        jobUpdater = new JenkinsJobUpdater(settings);
+    }
 
     @Test(expected = IOException.class)
     public void firtsWin() throws Throwable {
@@ -122,6 +147,78 @@ public class JenkinsJobUpdaterTest {
             Assert.assertNull(r);
         } finally {
             Assert.assertTrue(wasFinallyRun);
+        }
+    }
+
+    @Test
+    public void regenerateAllJDKProject() throws ManagementException, StorageException {
+        JDKProjectManager tpm = new JDKProjectManager(
+                cm,
+                folderHolder.reposRoot,
+                folderHolder.scriptsRoot
+        );
+        JenkinsCliWrapper.killCli();
+        try {
+            JobUpdateResults r1 = jobUpdater.regenerateAll(null, tpm,null);//create all
+            JobUpdateResults r2 = jobUpdater.regenerateAll(null, tpm, null);//re create all
+            Assert.assertEquals(0, r1.jobsRewritten.size());
+            Assert.assertNotEquals(0, r1.jobsCreated.size());
+            Assert.assertNotEquals(0, r2.jobsRewritten.size());
+            Assert.assertEquals(0, r2.jobsCreated.size());
+        } finally {
+            JenkinsCliWrapper.reinitCli();
+        }
+    }
+
+    @Test
+    public void regenerateAllJDKProjectWithWhitelist() throws ManagementException, StorageException {
+        JDKProjectManager tpm = new JDKProjectManager(
+                cm,
+                folderHolder.reposRoot,
+                folderHolder.scriptsRoot
+        );
+        JenkinsCliWrapper.killCli();
+        try {
+            JobUpdateResults r1 = jobUpdater.regenerateAll(null, tpm, "somethingNotExisting");//create nothing
+            JobUpdateResults r2 = jobUpdater.regenerateAll(null, tpm, null);//re create all
+            Assert.assertEquals(0, r1.jobsRewritten.size());
+            Assert.assertEquals(0, r1.jobsCreated.size());
+            Assert.assertEquals(0, r2.jobsRewritten.size());
+            Assert.assertNotEquals(0, r2.jobsCreated.size());
+        } finally {
+            JenkinsCliWrapper.reinitCli();
+        }
+    }
+
+    @Test
+    public void regenerateAllJDKTestProject() throws ManagementException, StorageException {
+        JDKTestProjectManager tpm = new JDKTestProjectManager(cm.getJdkTestProjectStorage());
+        JenkinsCliWrapper.killCli();
+        try {
+            JobUpdateResults r1 = jobUpdater.regenerateAll(null, tpm, null);//create all
+            JobUpdateResults r2 = jobUpdater.regenerateAll(null, tpm, null);//re create all
+            Assert.assertEquals(0, r1.jobsRewritten.size());
+            Assert.assertEquals(5, r1.jobsCreated.size());
+            Assert.assertEquals(5, r2.jobsRewritten.size());
+            Assert.assertEquals(0, r2.jobsCreated.size());
+        } finally {
+            JenkinsCliWrapper.reinitCli();
+        }
+    }
+
+    @Test
+    public void regenerateAllJDKTestProjectWithWhitelist() throws ManagementException, StorageException {
+        JDKTestProjectManager tpm = new JDKTestProjectManager(cm.getJdkTestProjectStorage());
+        JenkinsCliWrapper.killCli();
+        try {
+            JobUpdateResults r1 = jobUpdater.regenerateAll(null, tpm, "tck-.*");//create tck
+            JobUpdateResults r2 = jobUpdater.regenerateAll(null, tpm, null);//re create rest
+            Assert.assertEquals(0, r1.jobsRewritten.size());
+            Assert.assertEquals(4, r1.jobsCreated.size());
+            Assert.assertEquals(4, r2.jobsRewritten.size());
+            Assert.assertEquals(1,  r2.jobsCreated.size());
+        } finally {
+            JenkinsCliWrapper.reinitCli();
         }
     }
 }
